XRDDEV-371 both api key and session auth use (dummy) GrantedAuthorityMapper to register granted permissions in addition to roles
