+++ src/proxy-ui-api/frontend/vue.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   devServer: {
-    proxy: 'http://docker-rest-ui-ss.local:8020' // this needs to be parametrized (address where backend runs)
+    proxy: 'http://localhost:8020' // this needs to be parametrized (address where backend runs)
   },
 }
