+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/auth/ApiKeyAuthenticationManager.java
@@ -32,14 +32,8 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.web.authentication.preauth.PreAuthenticatedAuthenticationToken;
 import org.springframework.stereotype.Component;
-
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * AuthenticationManager which expects Authentication.principal to be
@@ -57,7 +51,7 @@ public class ApiKeyAuthenticationManager implements AuthenticationManager {
     private AuthenticationHeaderDecoder authenticationHeaderDecoder;
 
     @Autowired
-    private PermissionMapper permissionMapper;
+    private GrantedAuthorityMapper permissionMapper;
 
     @Override
     public Authentication authenticate(Authentication authentication) throws AuthenticationException {
@@ -70,33 +64,8 @@ public class ApiKeyAuthenticationManager implements AuthenticationManager {
         PreAuthenticatedAuthenticationToken authenticationWithGrants =
                 new PreAuthenticatedAuthenticationToken(authentication.getPrincipal(),
                         authentication.getCredentials(),
-                        getAllGrants(key.getRoles()));
+                        permissionMapper.getAuthorities(key.getRoles()));
         log.debug("authentication: {}", authenticationWithGrants);
         return authenticationWithGrants;
-    }
-
-    /**
-     * Create set of grants, consisting of roles and permissions
-     * @param rolenames
-     * @return
-     */
-    private Set<SimpleGrantedAuthority> getAllGrants(Collection<String> rolenames) {
-        Set<SimpleGrantedAuthority> grants = new HashSet<>();
-        grants.addAll(getRoleGrants(rolenames));
-        grants.addAll(getPermissionGrants(rolenames));
-        return grants;
-    }
-
-    private Set<SimpleGrantedAuthority> getPermissionGrants(Collection<String> rolenames) {
-        return permissionMapper.getPermissions(rolenames)
-                .stream()
-                .map(name -> new SimpleGrantedAuthority(name.toUpperCase()))
-                .collect(Collectors.toSet());
-    }
-
-    private Set<SimpleGrantedAuthority> getRoleGrants(Collection<String> rolenames) {
-        return rolenames.stream()
-                .map(name -> new SimpleGrantedAuthority("ROLE_" + name.toUpperCase()))
-                .collect(Collectors.toSet());
     }
 }
